% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATCCoding}
\alias{ATCCoding}
\title{Table of ATC codes of all categoric levels with german labels}
\format{
\subsection{'ATCCoding'}{

Tibble
\describe{
\item{Code}{Full ATCCode}
\item{Substance}{}
\item{Category1Code}{}
\item{Category1}{}
\item{Category3Code}{}
\item{Category3}{}
\item{Category4Code}{}
\item{Category4}{}
\item{Category5Code}{}
\item{Category5}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
ATCCoding
}
\description{
A tibble
}
\keyword{datasets}
