% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetECDF.R
\name{GetECDF}
\alias{GetECDF}
\title{GetECDF}
\usage{
GetECDF(DataFrame, MetricFeature, GroupingFeature = NULL, na.rm = FALSE)
}
\arguments{
\item{DataFrame}{Data frame that contains the metric feature}

\item{MetricFeature}{Metric feature}

\item{GroupingFeature}{Grouping feature}

\item{na.rm}{Logical; whether NA values should be removed before estimating the CDF}
}
\value{
An ECDF object
}
\description{
Get Empirical Cumulative Distribution Function (ECDF)
}
