% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMatchingGrade.R
\name{GetMatchingGrade}
\alias{GetMatchingGrade}
\title{GetMatchingGrade}
\usage{
GetMatchingGrade(VectorA, VectorB)
}
\arguments{
\item{VectorA}{\code{vector}}

\item{VectorB}{\code{vector}}
}
\value{
\code{double} - The proportion of matching elements
}
\description{
Calculates the proportion of matching elements between two vectors
}
\author{
Bastian Reiter
}
