% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNGrams.R
\name{GetNGrams}
\alias{GetNGrams}
\title{GetNGrams}
\usage{
GetNGrams(String, n = 1)
}
\arguments{
\item{String}{\code{string}}

\item{n}{\code{integer} - Length of tokens - Default: 1}
}
\value{
\code{vector} of n-grams
}
\description{
Get n-gram tokens from a character string
}
