% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleQuantiles.R
\name{GetSampleQuantiles}
\alias{GetSampleQuantiles}
\title{GetSampleQuantiles}
\usage{
GetSampleQuantiles(
  DataFrame,
  MetricFeature,
  GroupingFeature = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{DataFrame}{Data Frame that contains metric feature}

\item{MetricFeature}{Metric feature}

\item{GroupingFeature}{Grouping feature}

\item{na.rm}{Logical; whether NA values should be removed before calculating quantiles}
}
\value{
A tibble containing quantiles
}
\description{
Get key quantiles of a metric vector
}
