% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleStatistics.R
\name{GetSampleStatistics}
\alias{GetSampleStatistics}
\title{GetSampleStatistics}
\usage{
GetSampleStatistics(
  DataFrame,
  MetricFeature,
  GroupingFeature = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{DataFrame}{Data frame that contains the metric feature}

\item{MetricFeature}{Metric feature}

\item{GroupingFeature}{Grouping feature}

\item{na.rm}{Logical; whether NA values should be removed before calculating statistics}
}
\value{
A tibble containing the following statistics: sample size (N), minimum, maximum, Q1, Median, Q3, MAD, mean, standard deviation
}
\description{
Get important sample statistics for a feature of a data frame
}
