% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetValidityGrade.R
\name{GetValidityGrade}
\alias{GetValidityGrade}
\title{GetValidityGrade}
\usage{
GetValidityGrade(
  Vector,
  ValidValues = NULL,
  ValidRange.Min = NULL,
  ValidRange.Max = NULL
)
}
\arguments{
\item{Vector}{\code{vector} - Containing original values}

\item{ValidValues}{\code{vector} - A set of valid values}

\item{ValidRange.Min}{\code{double}}

\item{ValidRange.Max}{\code{double}}
}
\value{
\code{double} - The proportion of valid elements
}
\description{
Calculates the proportion of valid elements in a given vector.
}
\author{
Bastian Reiter
}
