% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_Tinker.R
\name{ggtheme_Tinker}
\alias{ggtheme_Tinker}
\title{ggtheme_Tinker}
\usage{
ggtheme_Tinker(
  ...,
  BaseSize = 11,
  LegendPosition = "right",
  SizeFactorPlotTitle = 1.67,
  SizeFactorPlotSubtitle = 1.33,
  SizeFactorPlotCaption = 1,
  SizeFactorAxisLabels = 1.33,
  SizeFactorTickLabels_x = 1,
  SizeFactorTickLabels_y = 1,
  SizeFactorLegendLabels = 1,
  SizeFactorFacetLabels = 1.33
)
}
\arguments{
\item{BaseSize}{Font size that relatively determines all other sizes}

\item{LegendPosition}{Position of legend ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{SizeFactorFacetLabels}{}
}
\description{
Custom ggplot2 theme, based on ggplot2::theme()
}
