% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertTypo.R
\name{InsertTypo}
\alias{InsertTypo}
\title{InsertTypo}
\usage{
InsertTypo(String, Type = "random")
}
\arguments{
\item{String}{\code{string}}

\item{Type}{\code{string} - Refers to the type of typo, one of the following:
\itemize{ \item 'double' - Clone a random letter and insert it
\item 'add' - Add a keyboard-neighboring letter to a random letter
\item 'replace' - Replace a random letter with a keyboard-neighboring letter
\item 'random' - Choose randomly one of the above (Default) }}
}
\value{
\code{string} with inserted typo
}
\description{
Insert a random typo into a string
}
\author{
Bastian Reiter
}
