% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDirty.R
\name{MakeDirty}
\alias{MakeDirty}
\title{MakeDirty}
\usage{
MakeDirty(Vector, AffectedProportion = 0.3, ReturnModificationTibble = FALSE)
}
\arguments{
\item{Vector}{\code{character vector} - 'Clean' string data}

\item{AffectedProportion}{\code{double} - Proportion of vector affected by noise - Default: 0.3}

\item{ReturnModificationTibble}{\code{logical} - Optionally return a \code{tibble} containing various modifications for the random selection of affected original values - Default: \code{FALSE}}
}
\value{
Either
\itemize{\item A \code{vector} of 'dirty' strings (default), or
\item If argument 'ReturnModificationTibble' is set \code{TRUE}, a \code{tibble} containing various modifications performed on the random selection of original values }
}
\description{
Introduce a variety of noise into clean text data, including:
\itemize{\item Typos
\item Truncation
\item Addition
\item First case flip
\item Total case flip}
}
\author{
Bastian Reiter
}
