% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePlot_Column.R
\name{MakePlot_Column}
\alias{MakePlot_Column}
\title{MakeColumn_Plot}
\usage{
MakePlot_Column(
  DataFrame,
  XFeature,
  XFeatureSpecs = NULL,
  XFeatureAdditionalMapping = "none",
  YFeature,
  GroupingFeature = NULL,
  GroupingSpecs = NULL,
  GroupingPosition = ggplot2::position_stack(),
  GroupingMapping = "fill",
  FacetFeature = NULL,
  FacetSpecs = NULL,
  FacetMapping = "none",
  FacetArguments = list(),
  AxisType_y = "absolute",
  CoordFlip = FALSE,
  LegendPosition = "right",
  LegendShowFillGuide = TRUE,
  AxisTitle_x = "",
  AxisTitle_y = "",
  TickLabelWidth_x = 10,
  Decimals = 0,
  ggTheme = function(...) TinkerLab::ggTheme_Tinker(...),
  ThemeArguments = list(),
  FillPalette = NULL,
  ColorPrimary = TinkerLab::Colors$Primary,
  AlphaPalette = NULL,
  ColumnWidth = 0.95,
  ...
)
}
\arguments{
\item{DataFrame}{Data frame that should have tidy format}

\item{XFeature}{Variable that contains X axis categories}

\item{XFeatureSpecs}{Optional: Character vector that determines selection and order of categories}

\item{XFeatureAdditionalMapping}{Optional: "fill" / "alpha" / "pattern" / "none" (default)}

\item{YFeature}{Variable that contains values / measurements}

\item{GroupingFeature}{Optional variable that contains group names}

\item{GroupingSpecs}{Optional: Character vector that determines three specifications of grouping: Selection and order of group levels (from top to bottom in stacked and left to right in dodged) and custom legend labeling (otherwise legend labels are taken from values of grouping variable)}

\item{GroupingPosition}{Optional: position_dodge() / position_fill() / position_stack() (default) / ...}

\item{GroupingMapping}{Optional: "alpha" / "pattern"}

\item{FacetFeature}{Optional variable for facet grouping}

\item{FacetSpecs}{Optional: Character vector that determines selection and order of facet levels}

\item{FacetMapping}{Optional: "fill" / "alpha" / "pattern" / "none" (default)}

\item{FacetArguments}{A list}

\item{AxisType_y}{Optional: "absolute" (default) / "proportional"}

\item{LegendPosition}{Optional: "top" / "bottom" / "left" / "none"}

\item{LegendShowFillGuide}{Optional logical: Show / hide legend guide of fill mapping}

\item{ThemeArguments}{Pass custom theme arguments as list object}

\item{FillPalette}{Custom color palette passed into function environment}

\item{...}{Anonymous additional arguments}
}
\value{
A ggplot2 object
}
\description{
MakeColumn_Plot
}
\author{
Bastian Reiter
}
