% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Res.CancerSurgery}
\alias{Res.CancerSurgery}
\title{Data on OPS codes in cancer surgery}
\format{
\subsection{'Res.CancerSurgery'}{

Tibble
\describe{
\item{ICD10Code.Short}{Three digit ICD-10 code}
\item{CancerTopographyDetail.ICD10}{Affected organ / topography as put forth by ICD-10}
\item{OPSCode.Short}{Four digit OPS code}
\item{Procedure}{German label of procedure coded in OPS}
\item{IsLikelyCancerRelated}{}
\item{IsLikelyCurativeIntention_Primary}{}
\item{IsLikelyCurativeIntention_Secondary}{}
\item{IsLikelySupportive_Direct}{}
\item{IsLikelySupportive_Indirect}{}
\item{IsInPlattform65cCatalogue}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
Res.CancerSurgery
}
\description{
A tibble
}
\keyword{datasets}
