% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Res.OPSCodes}
\alias{Res.OPSCodes}
\title{Data on OPS codes in general}
\format{
\subsection{'Res.OPSCodes'}{

Tibble
\describe{
\item{OPSVersion}{Year of OPS version}
\item{CodeLevel}{}
\item{IsTerminal}{}
\item{ChapterCode}{}
\item{FirstSubgroupCodeOfGroup}{}
\item{SubgroupCode}{}
\item{OPSCode}{}
\item{NeedsLocalization}{}
\item{Procedure}{}
\item{Procedure.Main}{}
\item{Procedure.Specification}{}
\item{Procedure.Subspecification}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
Res.OPSCodes
}
\description{
A tibble
}
\keyword{datasets}
