% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Res.P21.AdmissionCauses}
\alias{Res.P21.AdmissionCauses}
\title{Data on Admission cause coding in the context of P21 data}
\format{
\subsection{'Res.P21.AdmissionCauses'}{

Tibble
\describe{
\item{AdmissionCauseCode}{}
\item{AdmissionCause}{}
\item{AdmissionCauseLabel}{}
\item{AdmissionCauseLabel.Original}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
Res.P21.AdmissionCauses
}
\description{
A tibble
}
\keyword{datasets}
