% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Res.P21.Departments}
\alias{Res.P21.Departments}
\title{Data on Hospital Department coding in the context of P21 data}
\format{
\subsection{'Res.P21.Departments'}{

Tibble
\describe{
\item{DepartmentCode}{}
\item{DepartmentOriginalLabel}{}
\item{Department}{}
\item{OperatingSpecialty}{}
\item{Subspecialty}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
Res.P21.Departments
}
\description{
A tibble
}
\keyword{datasets}
