% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Res.P21.DischargeReasons}
\alias{Res.P21.DischargeReasons}
\title{Data on Discharge Reason coding in the context of P21 data}
\format{
\subsection{'Res.P21.DischargeReasons'}{

Tibble
\describe{
\item{DischargeReasonCode}{}
\item{DischargeReasonOriginalLabel}{}
\item{DischargeCategory}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
Res.P21.DischargeReasons
}
\description{
A tibble
}
\keyword{datasets}
