% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftValues.R
\name{ShiftValues}
\alias{ShiftValues}
\title{ShiftValues}
\usage{
ShiftValues(Vector, Lag)
}
\arguments{
\item{Vector}{A vector object}

\item{Lag}{Size of lag as integer}
}
\value{
A Vector
}
\description{
ShiftValues
}
