% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLToDataModel.R
\name{XMLToDataModel}
\alias{XMLToDataModel}
\title{XMLToDataModel}
\usage{
XMLToDataModel(
  XMLSchemaFilePath,
  TableNodeElement = "container",
  TableNameAttribute = "opal-table",
  ColumnNodeElement = "attribute",
  ColumnNameAttribute = "csv-column"
)
}
\arguments{
\item{ColumnNameAttribute}{}
}
\value{
A list of data frames
}
\description{
Make List of data frames describing a data model
}
