% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OPSCodes}
\alias{OPSCodes}
\title{Meta data on OPS codes in general}
\format{
\subsection{'OPSCodes'}{

Tibble
\describe{
\item{OPSVersion}{Year of OPS version}
\item{CodeLevel}{}
\item{IsTerminal}{}
\item{ChapterCode}{}
\item{FirstSubgroupCodeOfGroup}{}
\item{SubgroupCode}{}
\item{OPSCode}{}
\item{NeedsLocalization}{}
\item{Procedure}{}
\item{ProcedureMain}{}
\item{ProcedureSpecification}{}
\item{ProcedureSubspecification}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
OPSCodes
}
\description{
A tibble
}
\keyword{datasets}
