% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Res.ICDOMorphology}
\alias{Res.ICDOMorphology}
\title{Data on ICD-O-3 morphology codes}
\format{
\subsection{'Res.ICDOMorphology'}{

Tibble
\describe{
\item{ICDOMorphologyCode}{}
\item{ICDOMorphologyLabel}{}
\item{ICDOMorphologyGroupCode}{}
\item{ICDOMorphologyGroup}{}
\item{ICDOMorphologyGroup.Sub1}{}
\item{ICDOMorphologyGroup.Sub2}{}
\item{ICDOMorphologyBehaviorCode}{}
\item{ICDOMorphologyBehaviorLabel}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/TinkerLab/tree/main/Development/Data}
}
\usage{
Res.ICDOMorphology
}
\description{
A tibble
}
\keyword{datasets}
